lvl.py
This code is a Python script that calculates the number of years that have passed since the user's birthday (level_since_birthday), based on the current date and the user's birthday specified in the environment variable "birthday_yday" (year day).

The script first imports the "datetime" module and the "os" module, which allows the script to access the operating system's environment variables. It then retrieves the value of the "birthday_yday" variable from the environment and assigns it to the variable "birthday_yday".

The script then uses the "datetime" module to get the current date and time and assigns the current year to the variable "current_year" and the day of the year to the variable "day_of_year".

Then, it checks if the current day of the year is less than the birthday day of the year. If it is, it subtracts one from the current year, and adds 365 to the day of the year.

Then, it retrieves the value of the "birthday_year" variable from the environment and assigns it to the variable "birthday_year".

Finally, it calculates the number of years that have passed since the user's birthday by subtracting the "birthday_year" variable from the "current_year" variable, and adding a fraction that represents the number of days that have passed since the user's birthday day of the year, divided by the total number of days in a year, which is 365+ birthday_yday variable

And print the result.